refactor(runtime): refactoring of runtime, enabling caching for stores (enhanced rwasm/wasmtime performance by times of magnitude), fix race condition in inter-process-lock, refactor usage of runtime testing context, tiny refactoring of types, improve benches
