+++ crates/revm/src/executor.rs
@@ -6,7 +6,6 @@ use crate::{
     types::{SystemInterruptionInputs, SystemInterruptionOutcome},
     ExecutionResult, NextAction,
 };
-use core::cell::RefCell;
 use fluentbase_runtime::{
     instruction::{exec::SyscallExec, resume::SyscallResume},
     RuntimeContext,
@@ -177,28 +176,28 @@ fn execute_rwasm_frame<CTX: ContextTr, INSP: Inspector<CTX>>(
     };
     let bytecode_hash = BytecodeOrHash::Bytecode {
         address: effective_bytecode_address,
-        rwasm_module: rwasm_bytecode,
-        code_hash: rwasm_code_hash,
+        bytecode: rwasm_bytecode,
+        hash: rwasm_code_hash,
     };
 
-    // fuel limit we denominate later to gas
+    // Fuel limit we denominate later to gas
     let fuel_limit = interpreter
         .gas
         .remaining()
         .checked_mul(FUEL_DENOM_RATE)
         .unwrap_or(u64::MAX);
 
+    // Check whenever bytecode has self-managed gas,
+    // it's possible only for system precompiled contracts, like EVM, WASM, etc.
     let is_gas_free = fluentbase_sdk::is_system_precompile(&effective_bytecode_address);
 
-    // execute function
-    let mut runtime_context = RuntimeContext::root();
+    // Execute function
+    let mut runtime_context = RuntimeContext::default();
     if is_gas_free {
-        runtime_context = runtime_context.without_fuel();
+        runtime_context = runtime_context.with_disabled_fuel();
     }
-    let runtime_context = RefCell::new(runtime_context);
-
     let (fuel_consumed, fuel_refunded, exit_code) = SyscallExec::fn_impl(
-        &mut runtime_context.borrow_mut(),
+        &mut runtime_context,
         bytecode_hash,
         BytesOrRef::Bytes(context_input.into()),
         fuel_limit,
@@ -213,7 +212,7 @@ fn execute_rwasm_frame<CTX: ContextTr, INSP: Inspector<CTX>>(
 
     // extract return data from the execution context
     let return_data: Bytes;
-    return_data = runtime_context.borrow_mut().take_return_data().into();
+    return_data = runtime_context.execution_result.return_data.into();
     let gas = interpreter.gas;
 
     process_exec_result(
@@ -275,13 +274,12 @@ fn execute_rwasm_resume<CTX: ContextTr, INSP: Inspector<CTX>>(
         _ => ExitCode::UnknownError,
     };
 
-    let mut runtime_context = RuntimeContext::root();
+    let mut runtime_context = RuntimeContext::default();
     if inputs.is_gas_free {
-        runtime_context = runtime_context.without_fuel();
+        runtime_context = runtime_context.with_disabled_fuel();
     }
-    let runtime_context = RefCell::new(runtime_context);
     let (fuel_consumed, fuel_refunded, exit_code) = SyscallResume::fn_impl(
-        &mut runtime_context.borrow_mut(),
+        &mut runtime_context,
         inputs.call_id,
         result.output.as_ref(),
         exit_code.into_i32(),
@@ -289,7 +287,7 @@ fn execute_rwasm_resume<CTX: ContextTr, INSP: Inspector<CTX>>(
         fuel_refunded,
         inputs.syscall_params.fuel16_ptr,
     );
-    let return_data: Bytes = runtime_context.borrow_mut().take_return_data().into();
+    let return_data: Bytes = runtime_context.execution_result.return_data.into();
 
     // if we're free from paying gas,
     // then just take the previous gas value and don't charge anything
