+++ crates/revm/src/lib.rs
@@ -22,6 +22,5 @@ pub use evm::RwasmEvm;
 pub use handler::*;
 pub use precompiles::*;
 pub use result::*;
-pub use spec::*;
-
 use rwasm as _;
+pub use spec::*;
