+++ crates/runtime/src/context.rs
@@ -1,44 +1,57 @@
-use crate::{instruction::exec::SysExecResumable, ExecutionResult};
-use fluentbase_types::Bytes;
-use std::mem::take;
+use crate::{instruction::exec::InterruptionHolder, ExecutionResult};
+use fluentbase_types::{Bytes, CALL_DEPTH_ROOT, STATE_MAIN};
 
 #[derive(Debug, Clone)]
 pub struct RuntimeContext {
-    // context inputs
-    pub(crate) disable_fuel: bool,
-    pub(crate) state: u32,
-    pub(crate) call_depth: u32,
-    pub(crate) input: Bytes,
-    // context outputs
-    pub(crate) execution_result: ExecutionResult,
-    pub(crate) resumable_context: Option<SysExecResumable>,
+    /// A fuel limit for this call. If it's set to `None` then fuel is disabled.
+    pub fuel_limit: u64,
+    /// Do we disable fuel for this call?
+    /// With this flag enabled,
+    /// we ignore fuel checks inside VMs and the only way to manage fuel is by using builtins:
+    /// `charge_fuel` and `charge_fuel_manually`
+    pub disable_fuel: bool,
+    /// An execution state.
+    /// Usually we use values like `STATE_MAIN (0)`
+    /// and `STATE_DEPLOY (1)` to identify application execution state: deployment or execution.
+    /// Theoretically, other states can be used, but we don't use it right now.
+    pub state: u32,
+    /// A call depth for this execution phrase.
+    /// Once it reached `MAX_CALL_DEPTH (1024)`, then execution halts.
+    /// By checking call depth to zero, we determine so-called root layer.
+    pub call_depth: u32,
+    /// An input for the application.
+    /// Just an input.
+    /// It already contains all required context params inside.
+    pub input: Bytes,
+    /// Execution result details, like exit code or output data.
+    pub execution_result: ExecutionResult,
+    /// An information about interrupted call.
+    pub resumable_context: Option<InterruptionHolder>,
 }
 
 impl Default for RuntimeContext {
     fn default() -> Self {
         Self {
-            state: 0,
-            input: Bytes::default(),
-            call_depth: 0,
-            execution_result: ExecutionResult::default(),
+            fuel_limit: 0,
             disable_fuel: false,
+            state: STATE_MAIN,
+            call_depth: CALL_DEPTH_ROOT,
+            input: Bytes::default(),
+            execution_result: ExecutionResult::default(),
             resumable_context: None,
         }
     }
 }
 
 impl RuntimeContext {
-    pub fn root() -> Self {
-        Self::default().with_call_depth(0)
+    pub fn with_fuel_limit(mut self, fuel_limit: u64) -> Self {
+        self.fuel_limit = fuel_limit;
+        self
     }
 
     pub fn with_input<I: Into<Bytes>>(mut self, input_data: I) -> Self {
         self.input = input_data.into();
         self
-    }
-
-    pub fn change_input(&mut self, input_data: Bytes) {
-        self.input = input_data;
     }
 
     pub fn with_state(mut self, state: u32) -> Self {
@@ -51,84 +64,16 @@ impl RuntimeContext {
         self
     }
 
-    pub fn with_disable_fuel(mut self, disable_fuel: bool) -> Self {
-        self.disable_fuel = disable_fuel;
-        self
-    }
-
-    pub fn without_fuel(mut self) -> Self {
+    pub fn with_disabled_fuel(mut self) -> Self {
         self.disable_fuel = true;
         self
-    }
-
-    pub fn depth(&self) -> u32 {
-        self.call_depth
-    }
-
-    pub fn exit_code(&self) -> i32 {
-        self.execution_result.exit_code
-    }
-
-    pub fn input(&self) -> &Bytes {
-        &self.input
-    }
-
-    pub fn input_size(&self) -> u32 {
-        self.input.len() as u32
-    }
-
-    pub fn output(&self) -> &Vec<u8> {
-        &self.execution_result.output
-    }
-
-    pub fn output_mut(&mut self) -> &mut Vec<u8> {
-        &mut self.execution_result.output
-    }
-
-    pub fn return_data(&self) -> &Vec<u8> {
-        &self.execution_result.return_data
-    }
-
-    pub fn into_return_data(self) -> Bytes {
-        self.execution_result.return_data.into()
-    }
-
-    pub fn take_return_data(&mut self) -> Vec<u8> {
-        take(&mut self.execution_result.return_data)
-    }
-
-    pub fn return_data_mut(&mut self) -> &mut Vec<u8> {
-        &mut self.execution_result.return_data
-    }
-
-    pub fn state(&self) -> u32 {
-        self.state
     }
 
     pub fn clear_output(&mut self) {
         self.execution_result.output.clear();
     }
 
-    // pub fn try_consume_fuel(&mut self, fuel: u64) -> Result<(), TrapCode> {
-    //     let consumed_fuel = self
-    //         .execution_result
-    //         .fuel_consumed
-    //         .checked_add(fuel)
-    //         .unwrap_or(u64::MAX);
-    //     if !self.disable_fuel {
-    //         if consumed_fuel > self.fuel_limit {
-    //             return Err(TrapCode::OutOfFuel);
-    //         }
-    //     }
-    //     self.execution_result.fuel_consumed = consumed_fuel;
-    //     Ok(())
-    // }
-
-    pub fn refund_fuel(&mut self, fuel: i64) {
-        self.execution_result.fuel_refunded += fuel;
+    pub fn is_fuel_disabled(&self) -> bool {
+        self.disable_fuel
     }
-
-    // pub fn remaining_fuel(&self) -> u64 {
-    //     self.fuel_limit - self.execution_result.fuel_consumed
-    // }
 }
