+++ crates/runtime/src/instruction/curve25519_edwards_decompress_validate.rs
@@ -1,4 +1,4 @@
-use crate::{utils::syscall_process_exit_code, RuntimeContext};
+use crate::{instruction::syscall_process_exit_code, RuntimeContext};
 use curve25519_dalek::{edwards::CompressedEdwardsY, EdwardsPoint};
 use fluentbase_types::ExitCode;
 use rwasm::{Store, TrapCode, TypedCaller, Value};
