+++ crates/runtime/src/instruction/curve25519_ristretto_decompress_validate.rs
@@ -1,4 +1,4 @@
-use crate::{utils::syscall_process_exit_code, RuntimeContext};
+use crate::{instruction::syscall_process_exit_code, RuntimeContext};
 use curve25519_dalek::{ristretto::CompressedRistretto, RistrettoPoint};
 use fluentbase_types::ExitCode;
 use rwasm::{Store, TrapCode, TypedCaller, Value};
