+++ crates/types/src/bytecode.rs
@@ -1,21 +1,5 @@
-use crate::KECCAK_EMPTY;
-use alloy_primitives::{Address, Bytes, B256};
-
-pub const EIP7702_SIG_LEN: usize = 2;
-/// rWASM binary format signature:
-/// - 0xef 0x00 - EIP-3540 compatible prefix
-/// - 0x52 - rWASM version number (equal to 'R')
-pub const EIP7702_SIG: [u8; EIP7702_SIG_LEN] = [0xef, 0x01];
-
-pub const WASM_SIG_LEN: usize = 4;
-/// WebAssembly signature (\00ASM)
-pub const WASM_SIG: [u8; WASM_SIG_LEN] = [0x00, 0x61, 0x73, 0x6d];
-
-pub const RWASM_SIG_LEN: usize = 2;
-/// rWASM binary format signature:
-/// - 0xef 0x00 - EIP-3540 compatible prefix
-/// - 0x52 - rWASM version number (equal to 'R')
-pub const RWASM_SIG: [u8; RWASM_SIG_LEN] = [0xef, 0x52];
+use crate::{Address, Bytes, B256, KECCAK_EMPTY, RWASM_SIG, RWASM_SIG_LEN, WASM_SIG, WASM_SIG_LEN};
+use core::fmt::Formatter;
 
 #[derive(Debug, Clone, Copy, PartialEq)]
 #[allow(non_camel_case_types)]
@@ -42,23 +26,30 @@ impl BytecodeType {
 #[derive(Clone, Debug)]
 pub enum BytecodeOrHash {
     Bytecode {
+        bytecode: Bytes,
+        hash: B256,
         address: Address,
-        rwasm_module: Bytes,
-        code_hash: B256,
     },
     Hash(B256),
 }
 
-#[derive(Clone, Debug)]
-pub enum BytesOrRef<'a> {
-    Bytes(Bytes),
-    Ref(&'a [u8]),
-}
-impl<'a> BytesOrRef<'a> {
-    pub fn into_bytes(self) -> Bytes {
+impl core::fmt::Display for BytecodeOrHash {
+    fn fmt(&self, f: &mut Formatter<'_>) -> core::fmt::Result {
         match self {
-            BytesOrRef::Bytes(bytes) => bytes,
-            BytesOrRef::Ref(slice) => Bytes::copy_from_slice(slice),
+            BytecodeOrHash::Bytecode {
+                address,
+                bytecode: rwasm_module,
+                hash: code_hash,
+            } => {
+                write!(
+                    f,
+                    "bytecode {}::{}::{}",
+                    address,
+                    rwasm_module.len(),
+                    code_hash
+                )
+            }
+            BytecodeOrHash::Hash(code_hash) => write!(f, "{}", code_hash),
         }
     }
 }
@@ -74,8 +65,8 @@ impl From<(Address, Bytes, B256)> for BytecodeOrHash {
     fn from(value: (Address, Bytes, B256)) -> Self {
         Self::Bytecode {
             address: value.0,
-            rwasm_module: value.1,
-            code_hash: value.2,
+            bytecode: value.1,
+            hash: value.2,
         }
     }
 }
@@ -85,8 +76,8 @@ impl Default for BytecodeOrHash {
     fn default() -> Self {
         Self::Bytecode {
             address: Address::ZERO,
-            rwasm_module: Bytes::new(),
-            code_hash: KECCAK_EMPTY,
+            bytecode: Bytes::new(),
+            hash: KECCAK_EMPTY,
         }
     }
 }
@@ -94,8 +85,25 @@ impl Default for BytecodeOrHash {
 impl BytecodeOrHash {
     pub fn code_hash(&self) -> B256 {
         match self {
-            BytecodeOrHash::Bytecode { code_hash, .. } => *code_hash,
+            BytecodeOrHash::Bytecode {
+                hash: code_hash, ..
+            } => *code_hash,
             BytecodeOrHash::Hash(hash) => *hash,
+        }
+    }
+}
+
+#[derive(Clone, Debug)]
+pub enum BytesOrRef<'a> {
+    Bytes(Bytes),
+    Ref(&'a [u8]),
+}
+
+impl<'a> BytesOrRef<'a> {
+    pub fn into_bytes(self) -> Bytes {
+        match self {
+            BytesOrRef::Bytes(bytes) => bytes,
+            BytesOrRef::Ref(slice) => Bytes::copy_from_slice(slice),
         }
     }
 }
