+++ crates/types/src/lib.rs
@@ -4,10 +4,12 @@
 
 extern crate alloc;
 extern crate core;
+
 pub extern crate rwasm as rwasm_core;
 
 mod address;
-mod bytecode_type;
+mod block_fuel;
+mod bytecode;
 mod context;
 mod curves;
 pub mod evm;
@@ -15,6 +17,7 @@ mod exit_code;
 pub mod genesis;
 pub mod hashes;
 pub mod helpers;
+mod import_linker;
 mod native_api;
 mod preimage;
 mod rwasm;
@@ -24,13 +27,15 @@ mod syscall;
 
 pub use address::*;
 pub use alloy_primitives::*;
-pub use bytecode_type::*;
+pub use block_fuel::*;
+pub use bytecode::*;
 pub use byteorder;
 pub use context::*;
 pub use curves::*;
 pub use exit_code::*;
 pub use genesis::*;
 pub use hashbrown::{hash_map, hash_set, HashMap, HashSet};
+pub use import_linker::*;
 pub use native_api::*;
 pub use preimage::*;
 pub use rwasm::*;
@@ -53,6 +58,8 @@ pub const SYSTEM_ADDRESS: Address = address!("0xffffffffffffffffffffffffffffffff
 
 pub const STATE_MAIN: u32 = 0;
 pub const STATE_DEPLOY: u32 = 1;
+
+pub const CALL_DEPTH_ROOT: u32 = 0;
 
 /// A chain id for Fluent Developer Preview, where value hex is equal to 0x5201 where:
 /// - 0x52 - is ASCII of R
@@ -101,6 +108,22 @@ pub const EVM_MAX_CODE_SIZE: usize = revm_primitives::eip170::MAX_CODE_SIZE;
 ///
 /// Limit of maximum initcode size is `2 * WASM_MAX_CODE_SIZE`.
 pub const EVM_MAX_INITCODE_SIZE: usize = 2 * EVM_MAX_CODE_SIZE;
+
+pub const EIP7702_SIG_LEN: usize = 2;
+/// rWASM binary format signature:
+/// - 0xef 0x00 - EIP-3540 compatible prefix
+/// - 0x52 - rWASM version number (equal to 'R')
+pub const EIP7702_SIG: [u8; EIP7702_SIG_LEN] = [0xef, 0x01];
+
+pub const WASM_SIG_LEN: usize = 4;
+/// WebAssembly signature (\00ASM)
+pub const WASM_SIG: [u8; WASM_SIG_LEN] = [0x00, 0x61, 0x73, 0x6d];
+
+pub const RWASM_SIG_LEN: usize = 2;
+/// rWASM binary format signature:
+/// - 0xef 0x00 - EIP-3540 compatible prefix
+/// - 0x52 - rWASM version number (equal to 'R')
+pub const RWASM_SIG: [u8; RWASM_SIG_LEN] = [0xef, 0x52];
 
 #[macro_export]
 macro_rules! bn254_add_common_impl {
