+++ e2e/Cargo.toml
@@ -88,7 +88,3 @@ harness = false
 [[bench]]
 name = "greeting"
 harness = false
-
-[[bench]]
-name = "multicall"
-harness = false
