+++ e2e/benches/greeting.rs
@@ -8,7 +8,33 @@ use hex_literal::hex;
 fn greeting_benches(c: &mut Criterion) {
     let mut group = c.benchmark_group("Greeting Contract Comparison");
 
-    // --- Benchmark 1: EVM "Hello World" Contract ---
+    // --- Benchmark 1: Original "Hello World" Contract ---
+    {
+        let mut ctx = EvmTestingContext::default().with_full_genesis();
+        ctx.disabled_rwasm = true;
+        const OWNER_ADDRESS: Address = Address::ZERO;
+        let contract_address = ctx.deploy_evm_tx(
+            OWNER_ADDRESS,
+            hex::decode(include_bytes!("../assets/HelloWorld.bin"))
+                .unwrap()
+                .into(),
+        );
+        let call_payload: Bytes = hex!("45773e4e").into();
+
+        group.bench_function("Original_Greeting", |b| {
+            b.iter(|| {
+                ctx.call_evm_tx(
+                    OWNER_ADDRESS,
+                    contract_address,
+                    call_payload.clone(),
+                    None,
+                    None,
+                );
+            });
+        });
+    }
+
+    // --- Benchmark 2: EVM "Hello World" Contract ---
     {
         let mut ctx = EvmTestingContext::default().with_full_genesis();
         const OWNER_ADDRESS: Address = Address::ZERO;
@@ -33,7 +59,7 @@ fn greeting_benches(c: &mut Criterion) {
         });
     }
 
-    // --- Benchmark 2: Wasm "Greeting" Contract ---
+    // --- Benchmark 3: Wasm "Greeting" Contract ---
     {
         let mut ctx = EvmTestingContext::default().with_full_genesis();
         const OWNER_ADDRESS: Address = Address::ZERO;
